docs: add Sphinx integration and stop using`dataclasses`
Stop using the built-in module `dataclasses` as there was no usage at all. The documentation related to `__init__` methods was moved to the one of the relative classes. Use "Parameters" instead of "Args". Add a simple script which automatically adds new objects to the source which is used by Sphinx to generate the documentation. Add `pyrubrum/assets/ as a submodule. Add the requirements needed to work with the documentation.