docs: automatically copy files from the main directory
Fix documentation errors related to indentation. Update `changelog_generator.py` and make it suitable for post-commit hooks. Remove the "Table of Contents" from CHANGELOG.md as it was not supported by ReStructuredText and Sphinx. The "utils" directory is no more, "tree" and "types" are now used instead. Exceptions are now supported. Update SECURITY.md and fix style for README.md. Update the requirements for documentation purposes.